代码/Epp-BackEnd/backend/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
azure-core==1.30.1
azure-storage-blob==12.19.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.5
distro==1.9.0
Django==4.2.11
django-common-helpers==0.9.2
django-cors-headers==4.3.1
django-cron==0.5.0
django-crontab==0.7.1
environs==9.5.0
exceptiongroup==1.2.1
faiss-cpu==1.8.0
filelock==3.13.4
frozenlist==1.4.1
fsspec==2024.3.1
GPUtil==1.4.0
grpcio==1.60.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.7
importlib_metadata==7.1.0
isodate==0.6.1
Jinja2==3.1.3
Markdown==3.3.6
MarkupSafe==2.1.5
marshmallow==3.21.1
minio==7.2.5
mpmath==1.3.0
multidict==6.0.5
mysqlclient==2.2.4
networkx==3.2.1
numpy==1.26.4
openai==0.28.0
packaging==24.0
pandas==2.2.2
pdfkit==0.6.1
pillow==10.3.0
protobuf==5.26.1
psutil==5.9.8
pyarrow==15.0.2
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.0
pydantic_core==2.18.1
pymilvus==2.4.0
PyMySQL==1.1.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
safetensors==0.4.3
scipy==1.13.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
sympy==1.12
tokenizers==0.15.2
torch==2.0.0
tqdm==4.66.2
transformers==4.39.2
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.9.0
urllib3==2.2.1
yarl==1.9.4
zipp==3.18.1
